requirements.txt
pytelegrambotapi==3.7.6
mysql-connector-python==8.0.23
PyYAML==5.4.1
